refactor: created folder for login and video endpoints and merged Client with YoutubeApi
